refactor: 폴더 경로 수정 (#89)
